refactor: seperate StarPoints component
